feature: move code, remove visit code, add read number task, add error record

+++ src/main/java/com/kazma233/blog/entity/article/Article.java
@@ -26,7 +26,7 @@ public class Article {
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
     private LocalDateTime updateTime;
 
-    private Integer readNum;
+    private Long readNum;
 
     private String archiveDate;
